refactor: 실행기 / 요청기 객체 성격에 맞게 분리
